requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
nemoguardrails>=0.10.0
openai>=2.0.0
python-dotenv>=1.0.0
